Revisit bug 1
Refactored code and reuploaded fix bug 1 with the use of unit testing
