+++ util.php
